requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pandas>=1.2.0
Pillow>=8.0.0
matplotlib>=3.4.0